agent/requirements.txt
openai-agents>=0.0.7
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
